+++ tests/2xx/246_set_result_code.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+. common.sh
+
+assert_func set_result_code $TEST_FAIL $TEST_EMPTY
+assert_func set_result_code $TEST_FAIL $TEST_EMPTY "BLA"
+assert_func set_result_code $TEST_OK $TEST_EMPTY "0"